Refactor to move custom css to its own class.
